changed echo feedback + removed hard coded bashrc

+++ src/sys_setup/autostart.sh
@@ -27,12 +27,16 @@ execute_script_on_login()
 disable_script_on_login()
 {
     path_to_scipt=$1
+    rc_file=$(~/.bashrc)
 
     if check_string_in_file "$path_to_scipt" "$rc_file";
     then
-        grep -v "$path_to_scipt" ~/.bashrc > ~/.bashrc_temp && mv ~/.bashrc_temp ~/.bashrc
+        grep -v "$path_to_scipt" "$rc_file" > ~/.bashrc_temp && mv ~/.bashrc_temp "$rc_file"
     fi
 
-    cd ~/.config/autostart/ || exit 1
-    rm org.gnome.Terminal.desktop
+    if check_file_existence "$HOME/.config/autostart/org.gnome.Terminal.desktop";
+    then
+        cd ~/.config/autostart/ || exit 1
+        rm org.gnome.Terminal.desktop
+    fi
 }
